refactor: coyote 와 catalina 의 역할에 따라 적절히 패키지 분리
