+++ tomcat/src/main/java/org/apache/catalina/SessionManager.java
@@ -1,10 +1,9 @@
-package org.apache.coyote.http11;
+package org.apache.catalina;
 
 import jakarta.servlet.http.HttpSession;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.UUID;
-import org.apache.catalina.Manager;
 
 public class SessionManager implements Manager {
