+++ tomcat/src/main/java/org/apache/coyote/http11/Http11Processor.java
@@ -14,6 +14,7 @@ import java.nio.file.Files;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.Optional;
+import org.apache.catalina.Cookie;
 import org.apache.coyote.Processor;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
